+++ lib/cocoapods/installer/target_installer.rb
@@ -204,6 +204,10 @@ module Pod
         end
       end
 
+      ENABLE_EXTERNAL_STRINGS_FILE_FLAG = {
+        :ios => Version.new('6'),
+        :osx => Version.new('10.8')
+      }
       # Creates a script that copies the resources to the bundle of the client
       # target.
       #
@@ -217,7 +221,9 @@ module Pod
         UI.message "- Generating copy resources script at #{UI.path(path)}" do
           resources = library.file_accessors.map { |accessor| accessor.resources.flatten.map {|res| project.relativize(res)} }.flatten
           resources << bridge_support_file if bridge_support_file
-          generator = Generator::CopyResourcesScript.new(resources)
+          platform_name = library.platform.name
+          reference_external_strings_file = library.platform.deployment_target >= ENABLE_EXTERNAL_STRINGS_FILE_FLAG[platform_name]
+          generator = Generator::CopyResourcesScript.new(resources, reference_external_strings_file)
           generator.save_as(path)
           add_file_to_support_group(path)
         end