+++ CHANGELOG.md
@@ -20,6 +20,8 @@ To install or update CocoaPods see this [guide](http://docs.cocoapods.org/guides
 
 * Inheriting `inhibit_warnings` per pod is now working
   [#1032](https://github.com/CocoaPods/CocoaPods/issues/1032)
+* Fix copy resources script for iOS < 6 and OS X < 10.8 by removing the `--reference-external-strings-file`
+  flag. [#1030](https://github.com/CocoaPods/CocoaPods/pull/1030)
 
 ## 0.19.1
 [CocoaPods](https://github.com/CocoaPods/CocoaPods/compare/0.19.0...0.19.1)